Extracting NewReminderInfoOptionsView to a separate view
